[SOL] Fixes required for Solang
* [SOL] Make lld thread-safe with llvm when used in-process

Every time Solang tries to link a web-assembly file in-process, the linker
re-inits llvm which is not thread-safe with the rest of solang.


* [SOL][BPF] Enable the _ExtInt extension on the BPF Target for Solana
