+++ lld/wasm/Driver.cpp
@@ -54,17 +54,6 @@ enum {
 #undef OPTION
 };
 
-// This function is called on startup. We need this for LTO since
-// LTO calls LLVM functions to compile bitcode files to native code.
-// Technically this can be delayed until we read bitcode files, but
-// we don't bother to do lazily because the initialization is fast.
-static void initLLVM() {
-  InitializeAllTargets();
-  InitializeAllTargetMCs();
-  InitializeAllAsmPrinters();
-  InitializeAllAsmParsers();
-}
-
 class LinkerDriver {
 public:
   void linkerMain(ArrayRef<const char *> argsArr);
@@ -94,7 +83,6 @@ bool link(ArrayRef<const char *> args, llvm::raw_ostream &stdoutOS,
   config = make<Configuration>();
   symtab = make<SymbolTable>();
 
-  initLLVM();
   LinkerDriver().linkerMain(args);
 
   return errorCount() == 0;
